requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Cython==3.0.11
Flask==3.0.3
future==1.0.0
fuzzywuzzy==0.18.0
gevent==24.2.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Levenshtein==0.26.0
MarkupSafe==2.1.5
mdc==1.2.1
numpy==2.1.1
packaging==24.1
pandas==2.2.3
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-Levenshtein==0.26.0
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.0
requests==2.32.3
six==1.16.0
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.0.3
